+++ app/containers/ContentAreaTop/messages.js
@@ -1,0 +1,25 @@
+/*
+ * Copyright © 2015-2017 Serenova, LLC. All rights reserved.
+ */
+
+/*
+ * ContentAreaTop Messages
+ *
+ * This contains all the text for the ContentAreaTop component.
+ */
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  wrapup: {
+    id: 'app.containers.ContentAreaTop.wrapup',
+    defaultMessage: 'Wrap Up',
+  },
+  wrapupOn: {
+    id: 'app.containers.ContentAreaTop.wrapupOn',
+    defaultMessage: 'Wrap Up On',
+  },
+  wrapupOff: {
+    id: 'app.containers.ContentAreaTop.wrapupOff',
+    defaultMessage: 'Wrap Up Off',
+  },
+});