+++ Changelog.md
@@ -2,6 +2,7 @@
 * CXV1-11110 - Make side panel collapse interaction specific. Only automatically open side panel on accepted interaction (when CRM is enabled) and script received for interaction
 * CXV1-9010 - Fix stats overflow for toolbar restricted dimensions on side panel collapse
 * CXV1-10779 - Fix resize for active phone panel and scripts only
+* CXV1-11423 - Replace Wrapup Toggle Switch in Toolbar with Options in Button Submenu
 
 ### [1.2.0] (2017-08-25)
 * CXV1-9006 - Toolbar view of scripts